rewrite to use data-access
