+++ package.json
@@ -6,7 +6,9 @@
         "Maarten Mulders <mthmulders@users.noreply.github.com>"
     ],
     "devDependencies": {
+        "@types/chai": "^4.0.6",
         "@types/chai-as-promised": "^7.1.0",
+        "@types/mocha": "^2.2.44",
         "@types/sinon": "^4.0.0",
         "@types/sinon-chai": "^2.7.29",
         "@types/source-map-support": "^0.4.0",
