--- packages/badge-service/Shared/models.js
@@ -1,9 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var Color;
-(function (Color) {
-    Color["Red"] = "red";
-    Color["Orange"] = "orange";
-    Color["Green"] = "green";
-})(Color = exports.Color || (exports.Color = {}));
-//# sourceMappingURL=models.js.map
