+++ packages/badge-service/package.json
@@ -8,18 +8,14 @@
     "clean": "rimraf src/**/*.js.map src/**/*.js",
     "prebuild": "npm run clean",
     "build": "tsc",
-    "test": "nyc --check-coverage --reporter=html --report-dir=reports/coverage --lines 80 --functions 80 --branches 75 mocha \"test/**/*.js\""
+    "test": "mocha \"test/**/*.js\""
   },
   "dependencies": {
-    "azure-storage": "^2.6.0",
-    "camelcase": "4.1.0",
     "js-sha512": "0.7.0",
-    "pascal-case": "2.0.1"
+    "stryker-dashboard-data-access": "file:../data-access"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "devDependencies": {
-    "@types/camelcase": "^4.1.0"
-  }
+  "devDependencies": {}
 }
