+++ packages/badge-service/scoreProcessor/function.json
@@ -2,7 +2,7 @@
  "disabled": false,
  "bindings": [
   {
-   "authLevel": "function",
+   "authLevel": "anonymous",
    "type": "httpTrigger",
    "direction": "in",
    "name": "req"
@@ -13,5 +13,5 @@
    "name": "res"
   }
  ],
- "scriptFile": "index.js"
+ "scriptFile": "scoreProcessor.js"
 }
