+++ packages/data-access/src/index.ts
@@ -2,3 +2,4 @@
 export { default as MutationScoreMapper } from './MutationScoreMapper';
 export { default as ProjectMapper } from './ProjectMapper';
 export { default as Project } from './models/Project';
+export { default as MutationScore } from './models/MutationScore';
